+++ fud/fud/stages/vivado/stage.py
@@ -49,7 +49,7 @@ class VivadoBaseStage(Stage):
             cmd = f"{self.remote_exec} {self.flags}"
 
         # Steps and schedule
-        local_tmpdir = self.setup_environment(verilog_path)
+        local_tmpdir = self.setup_environment(verilog_path, config)
         if use_ssh:
             remote_exec = RemoteExecution(self, config)
             remote_exec.import_libs()
@@ -66,7 +66,7 @@ class VivadoBaseStage(Stage):
 
         return local_tmpdir
 
-    def setup_environment(self, verilog_path):
+    def setup_environment(self, verilog_path, config):
         # Step 1: Make a new temporary directory
         @self.step()
         def mktmp() -> SourceType.Directory:
@@ -82,7 +82,7 @@ class VivadoBaseStage(Stage):
             """
             Copy device files into tmpdir.
             """
-            for f in self.device_files():
+            for f in self.device_files(config):
                 shutil.copy(f, tmpdir.name)
             shutil.copy(str(verilog_path), f"{tmpdir.name}/{self.target_name}")
